Added specializers example
